reorder code: remove from mod
