+++ core/lib/types/src/lib.rs
@@ -28,6 +28,7 @@ pub mod block;
 pub mod circuit;
 pub mod commitment;
 pub mod contract_verification_api;
+pub mod debug_flat_call;
 pub mod event;
 pub mod fee;
 pub mod fee_model;
