+++ core/lib/zksync_core/src/api_server/web3/backend_jsonrpsee/namespaces/debug.rs
@@ -1,5 +1,6 @@
 use zksync_types::{
-    api::{BlockId, BlockNumber, DebugCall, DebugCallFlat, ResultDebugCall, TracerConfig},
+    api::{BlockId, BlockNumber, DebugCall, ResultDebugCall, TracerConfig},
+    debug_flat_call::DebugCallFlat,
     transaction_request::CallRequest,
     H256,
 };
